Move TransactionError into the SDK
