+++ programs/failure/tests/failure.rs
@@ -1,11 +1,9 @@
 use solana_runtime::bank::Bank;
-use solana_runtime::bank::TransactionError;
 use solana_runtime::loader_utils::load_program;
-use solana_runtime::runtime::InstructionError;
 use solana_sdk::genesis_block::GenesisBlock;
 use solana_sdk::native_loader;
 use solana_sdk::native_program::ProgramError;
-use solana_sdk::transaction::Transaction;
+use solana_sdk::transaction::{InstructionError, Transaction, TransactionError};
 
 #[test]
 fn test_program_native_failure() {
