+++ runtime/tests/system.rs
@@ -1,10 +1,10 @@
-use solana_runtime::bank::{Bank, TransactionError};
-use solana_runtime::runtime::InstructionError;
+use solana_runtime::bank::Bank;
 use solana_sdk::genesis_block::GenesisBlock;
 use solana_sdk::native_program::ProgramError;
 use solana_sdk::signature::{Keypair, KeypairUtil};
 use solana_sdk::system_instruction::SystemInstruction;
 use solana_sdk::system_program;
+use solana_sdk::transaction::{InstructionError, TransactionError};
 use solana_sdk::transaction_builder::{BuilderInstruction, TransactionBuilder};
 
 struct SystemBank<'a> {
