+++ sdk/src/system_instruction.rs
@@ -2,6 +2,13 @@ use crate::pubkey::Pubkey;
 use crate::system_program;
 use crate::transaction_builder::BuilderInstruction;
 
+#[derive(Serialize, Debug, Clone, PartialEq)]
+pub enum SystemError {
+    AccountAlreadyInUse,
+    ResultWithNegativeLamports,
+    SourceNotSystemAccount,
+}
+
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub enum SystemInstruction {
     /// Create a new account
